Fix static analysis errors
Tested using valgrind, scanbuild and cppcheck.

* Found that mhash has a 4GB limit, imposed a hard limit on ms3_put()
* Fixed some leaks on error conditions
* Other minor issues

+++ src/response.c
@@ -36,7 +36,7 @@ uint8_t parse_list_response(const char *data, size_t length, ms3_list_st **list)
   time_t tout= 0;
   ms3_list_st *nextptr= NULL, *lastptr= NULL;
 
-  doc= xmlReadMemory(data, length, "noname.xml", NULL, 0);
+  doc= xmlReadMemory(data, (int)length, "noname.xml", NULL, 0);
   if (not doc)
   {
     return MS3_ERR_RESPONSE_PARSE;
